% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobel_api.R
\name{get_nobel_prizes}
\alias{get_nobel_prizes}
\title{Get Nobel Prize basic information}
\usage{
get_nobel_prizes(category = NULL, year = NULL)
}
\arguments{
\item{category}{character, optional. One of "Physics", "Chemistry", "Medicine", "Literature", "Peace", "Economics"}

\item{year}{integer, optional. Prize year.}
}
\value{
tibble
}
\description{
Get Nobel prizes data from official Nobel Prize API.
}
